requirements.txt
pandas
snowflake-connector-python
dbt-core==1.10.6
dbt-snowflake==1.10.0
